pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = ""
authors = [
    {name = "dimash",email = "arystambekdimash005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false